radiantplanks_backend/requirements.txt
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.2
django-cors-headers==4.5.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
fonttools==4.54.1
h11==0.14.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
loguru==0.7.2
lxml==5.3.0
nest-asyncio==1.6.0
nulltype==2.3.1
numpy==2.1.3
openpyxl==3.1.5
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyee==11.1.1
pyHanko==0.25.2
pyhanko-certvalidator==0.26.4
PyJWT==2.9.0
pypdf==5.1.0
pyphen==0.15.0
pyppeteer==2.0.0
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==1.26.20
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
win32-setctime==1.1.0
wsproto==1.2.0
xhtml2pdf==0.2.16
xlrd==2.0.1
zipp==3.21.0
zopfli==0.2.3.post1
